More GPU compile fixes: Addes CUDA as an explicit optional dependency.
