+++ mundy/agents/cmake/Dependencies.cmake
@@ -3,8 +3,8 @@ tribits_package_define_dependencies(
   LIB_OPTIONAL_PACKAGES
   TEST_REQUIRED_PACKAGES
   TEST_OPTIONAL_PACKAGES
-  LIB_REQUIRED_TPLS STK Teuchos fmt
+  LIB_REQUIRED_TPLS Kokkos STK Teuchos fmt
   LIB_OPTIONAL_TPLS MPI
   TEST_REQUIRED_TPLS GTest
-  TEST_OPTIONAL_TPLS
+  TEST_OPTIONAL_TPLS CUDA
   )
