+++ mundy/alens/cmake/Dependencies.cmake
@@ -4,7 +4,7 @@ tribits_package_define_dependencies(
   TEST_REQUIRED_PACKAGES
   TEST_OPTIONAL_PACKAGES
   LIB_REQUIRED_TPLS STK Teuchos Kokkos KokkosKernels fmt
-  LIB_OPTIONAL_TPLS MPI STKFMM PVFMM
+  LIB_OPTIONAL_TPLS MPI STKFMM PVFMM CUDA
   TEST_REQUIRED_TPLS GTest OpenRAND
   TEST_OPTIONAL_TPLS
   )
