+++ mundy/core/cmake/Dependencies.cmake
@@ -4,7 +4,7 @@ tribits_package_define_dependencies(
   TEST_REQUIRED_PACKAGES
   TEST_OPTIONAL_PACKAGES
   LIB_REQUIRED_TPLS Kokkos fmt
-  LIB_OPTIONAL_TPLS Teuchos
+  LIB_OPTIONAL_TPLS Teuchos MPI
   TEST_REQUIRED_TPLS GTest
-  TEST_OPTIONAL_TPLS
+  TEST_OPTIONAL_TPLS CUDA
   )
