+++ scrap/motion/cmake/Dependencies.cmake
@@ -3,7 +3,7 @@ tribits_package_define_dependencies(
   LIB_OPTIONAL_PACKAGES 
   TEST_REQUIRED_PACKAGES
   TEST_OPTIONAL_PACKAGES
-  LIB_REQUIRED_TPLS STK Teuchos fmt
+  LIB_REQUIRED_TPLS Kokkos STK Teuchos fmt
   LIB_OPTIONAL_TPLS MPI
   TEST_REQUIRED_TPLS GTest
   TEST_OPTIONAL_TPLS
